Squashed commit of the following:

commit 7a89525338dc1f48ff2d40cf1bd533e6c992fd22
Author: Rafał Słota <roxeon@gmail.com>
Date:   Wed Nov 20 11:34:49 2013 +0100

    VFS-244: fix sed compatibility

commit a03f4f133b67b7a6c8c4de845e5a5fe0b37f37e0
Author: Rafał Słota <roxeon@gmail.com>
Date:   Wed Nov 20 10:27:21 2013 +0100

    VFS-244: gen_dev: default cookie value

commit ac5df22e904197dcfe38e3bac9d227afc90a88f5
Author: Rafał Słota <roxeon@gmail.com>
Date:   Wed Nov 20 04:13:09 2013 +0100

    VFS-244: some bash scripts fixes

commit 72ddbe8ab2e482450697aaaa09aaf181f5ef905c
Author: Rafał Słota <roxeon@gmail.com>
Date:   Wed Nov 20 03:35:36 2013 +0100

    VFS-244: compile fix

commit 8c231e33209af052793e3e7df10f3e624ddb1c0f
Author: Rafał Słota <roxeon@gmail.com>
Date:   Wed Nov 20 03:11:42 2013 +0100

    VFS-181: revert helpers & enable debug in ct tests

commit 28e66a580eed6199cf50da274a37a9374a5e9f8e
Author: Rafał Słota <roxeon@gmail.com>
Date:   Wed Nov 20 02:51:06 2013 +0100

    VFS-181

commit 47aaf0632017a61d395d466c12c5e3e78e1ce7dd
Author: Rafał Słota <roxeon@gmail.com>
Date:   Wed Nov 20 02:19:05 2013 +0100

    VFS-244: compile fix

commit 7b88dc8c9e9f9a5ca61bcc3e7e103e107cb24154
Author: Rafał Słota <roxeon@gmail.com>
Date:   Wed Nov 20 00:30:38 2013 +0100

    VFS-181: some linkage fixes

commit 80aef9d6a533bafffc5caf28a188d212dc14a3ab
Author: Rafał Słota <roxeon@gmail.com>
Date:   Tue Nov 19 22:00:16 2013 +0100

    VFS-181: deps update && gen_config fix

commit d03bc82706d89bc93b6600f90ef377f43f97aa9d
Author: Rafał Słota <roxeon@gmail.com>
Date:   Tue Nov 19 20:58:28 2013 +0100

    VFS-244: compile fix

commit 1d1cdf5f774555e8cef5b8a1b710deaa3ebcad23
Author: Rafał Słota <roxeon@gmail.com>
Date:   Tue Nov 19 18:59:23 2013 +0100

    VFS-244: delete shares associated with deleted file

commit 64c1afabfc0ea1852e76db12ba157e1f8b5a10c7
Merge: 364743f 7ce0cf7
Author: Rafał Słota <roxeon@gmail.com>
Date:   Tue Nov 19 18:17:22 2013 +0100

    Merge branch 'master' into VFS-181

commit 364743f12e614796ee9a542c21047c3fdce02023
Author: Rafal Slota <roxeon@gmail.com>
Date:   Fri Oct 4 18:21:26 2013 +0200

    VFS-181: delete unused file_descriptors

+++ c_src/link_shared.sh
@@ -26,9 +26,37 @@ rm -f $BIN_DIR/lib${LIB_NAME}*
 
 for libfile in $@
 do
-    for file in $(ldd $BIN_DIR/$libfile | grep -v "$BIN_DIR/" | grep '=>' | awk -F'=>' '{print $2}' | awk -F' ' '{print $1}' | grep lib${LIB_NAME})
+    if [ ! -e $BIN_DIR/$libfile ]; then
+        continue;
+    fi
+
+
+    if [ "`uname -s`" = "Darwin" ]; then
+        libs_list=$(otool -L $BIN_DIR/$libfile | grep -v "$BIN_DIR/" | grep -v "@loader_path" | grep -v ":" | awk -F' ' '{print $1}' | grep lib${LIB_NAME})
+    else
+        libs_list=$(ldd $BIN_DIR/$libfile | grep -v "$BIN_DIR/" | grep '=>' | awk -F'=>' '{print $2}' | awk -F' ' '{print $1}' | grep lib${LIB_NAME})
+    fi
+
+    for file in $libs_list
     do
+        # Copy and link shared library
         cp -L $file ./$BIN_DIR
-        ln -sf `basename $file` $BIN_DIR/`basename $(echo $file | sed 's/\.so\..*/\.so/')`
+        link="`basename $file`"
+        target="$BIN_DIR/`basename $(echo $file | sed 's/\.so\..*/\.so/' | sed 's/\.[0-9.]*\.dylib.*/\.dylib/')`"
+        if [ `basename $link` != `basename $target` ]; then
+            ln -sf $link $target
+        fi
+
+        chmod u+w ./$BIN_DIR/$link
+        chmod u+w ./$BIN_DIR/$libfile
+
+        # Change rpath on OSX
+        if [ "`uname -s`" = "Darwin" ]; then
+            install_name=`otool -D ./$BIN_DIR/$link | tail -1`
+            install_name_tool -id "@loader_path/$link" ./$BIN_DIR/$link
+            install_name_tool -id "@loader_path/$libfile" ./$BIN_DIR/$libfile
+            install_name_tool -change "$install_name" "@loader_path/$link" ./$BIN_DIR/$libfile
+        fi
+
     done
 done
